Revamp serializer to allow for better meta-data
Serializer has been totally re-written to allow for entry of data
besides float or int data (and besides the corny exception that got
around that restriction, but only on the final line). The result is a
steaming pile of seemingly arbitrary string parsing and comparison rules
that I am not very proud to have written, but that being said, its still
a step up from before lol.

Add ability to attach input labels to input neurons. This will be good
for making networks feel more rooted to the problem being addressed.

Add concept of a "NetworkString". This is the output from
`Network.Serialize`. The idea is that if you do not have a network
instance but you DO have its serialized representation, you can still
poke at it for relevant information. This will be accomplished through
receiver functions, but they haven't been added yet.

Add metadata to network. This is to expose some statistics about the
network and its intent to consumers.

All timestamps are now stored as RFC3339 strings.

Results from predictions and training sessions are now stored when they
result in successes, on the network under its metadata.

Update README for new API changes.

Rename `TrainingConfiguration.TrainingIterations` to just `Iterations`
to reduce redundancy.