Complete JSON serialization and deserialization process
Shift to network package against my better judgement because it makes working with these structs easier

Add test for ensuring the deserialization and serialization processes are inverses of each other

Add `Equals` operations on all network structs. The is effectively the same as `reflect.DeepEqual` but I wanted more fine grained control so I wrote it manually

Added `ConnectWith` and `ConnectNeurons` functions. `ConnectWith` accepts a set of connections to connect the layers with instead of creating new connections from scratch. `ConnectNeurons` uses the existing connections in the network but sets the neuron that the connection connects to to the neuron found in the previous layer

Update run configurations