Fix gzip decompression
The problem stemmed from the fact that `io.Reader.Read` will only read up to `len(b)` bytes, and we are reusing the slice of bytes from the compressed data. Of course, compressed data is smaller than uncompressed data, hence `read` was only reading up to the length of the compressed data which caused the protocol buffer unmarshal to fail.