trainer.New now accepts an io.Writer for logging, so you can redirect the training progress to any useful location. Default is os.Stdout.
Update cmds to respect trainer.New contract update.

Fix a bug in proto translator in which the activation function name was not being put back into the deserialized network, just the function itself.

Fix a bug in SetConnectionWeights in which the error message was not using deeper indexes, which lead to the wrong message being reported.