Improve README
README now corresponds to actual code instead of old API code and contains rich examples and documentation.

Split and renamed `ResetForPass` into `ResetFromBatch` and `ResetFromPass` for it is clearer that this is a reset "from" a previous operation, not "for" an upcoming one. Also it is much more readable to have it in two separate functions than it is to pass in a boolean to trigger the two different operations. I find this clarity in readability to be more valuable than it was to eliminate repeated code.

`network.From` now validates that your neuron map is at least 2 layers long and each layer contains at least one neuron.

`network.IsFullyConnected` is added to report whether a network is fully connected across all layers or not.